build_scripts/pyproject.toml
[project]
name = "build"
version = "0.1.0"
description = ""
authors = [
    { name = "yiguo", email = "yiguo@example.com" }
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pyyaml (>=6.0.3,<7.0.0)",
    "requests (>=2.33.1,<3.0.0)",
    "typer (>=0.24.1,<0.25.0)"
]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
